FASTAPI/requirements.txt
fastapi[all]
sqlalchemy
databases
jinja2
passlib
python-jose
redis